% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnnpt.R
\name{bnnpt}
\alias{bnnpt}
\title{BNNPT (Bagging Nearest-Neighbor Prediction independence Test)}
\usage{
bnnpt(x, y, bags, mtry, perms)
}
\arguments{
\item{x}{A vector containing values of a continuous variable (X).}

\item{y}{A vector containing values of a continuous variable (Y).}

\item{bags}{the number of bagging.}

\item{mtry}{the number of randomly selected values from X.}

\item{perms}{the number of permutations to calculate permutation pvalues.}
}
\value{
pvalue The pvalue of BNNPT.
}
\description{
Bagging Nearest-Neighbor Prediction independence Test: an efficient method for nonlinear dependence of two continuous variables.
Testing dependence/correlation of two variables is one of the fundamental tasks in statistics. In this work, we proposed an efficient method for nonlinear dependence of two continuous variables (X and Y).
We addressed this research question by using BNNPT (Bagging Nearest-Neighbor Prediction independence Test).
In the BNNPT framework, we first used the value of X to construct a bagging neighborhood structure. We then obtained the out of bag estimator of Y based on the bagging neighborhood structure.
The square error was calculated to measure how well Y is predicted by X. Finally, a permutation test was applied to determine the significance of the observed square error.
}
\examples{

#Example1
set.seed(1234)
x<-rnorm(1000)
y<-rnorm(1000)
#Test whether X and Y are independent or not using BNNPT
library(BNNPT)
#Set the parameters: bags= 20, mtry = 10, perms = 100
t1<- Sys.time()
bnnpt(x,y,bags=20, mtry=10, perm=100)
t2<- Sys.time()
#Output calculation time of BNNPT
t2-t1

#Example2
#Generate random variables X and Y with sin function
set.seed(1234)
x<-rnorm(1000)
y<-sin(x)
#Test whether X and Y are independent or not using BNNPT
library(BNNPT)
t1<- Sys.time()
#Set the parameters: bags= 50, mtry = 10, perms = 100
bnnpt(x,y,bags=50, mtry=10, perm=100)
t2<- Sys.time()
#Output calculation time of BNNPT
t2-t1

}
\references{
Wang, Yi, Yi Li, Xiaoyu Liu, Weilin Pu, Xiaofeng Wang, Jiucun Wang, Momiao Xiong, Yin Yao Shugart, and Li Jin. "Bagging Nearest-Neighbor Prediction independence Test: an efficient method for nonlinear dependence of two continuous variables." Scientific reports 7, no. 1 (2017): 12736.
}
\author{
Yi Li, \email{liyistat@gmail.com}
}
